+++ fpdf/output.py
@@ -578,7 +578,7 @@ class OutputProducer:
                 # recalcTimestamp=False means that it doesn't modify the "modified" timestamp in head table
                 # if we leave recalcTimestamp=True the tests will break every time
                 fonttools_font = ttLib.TTFont(
-                    file=font["ttffile"], recalcTimestamp=False, fontNumber=0
+                    file=font["ttffile"], recalcTimestamp=False, fontNumber=0, lazy=True
                 )
 
                 # 1. get all glyphs in PDF