+++ test/end_to_end_legacy/__init__.py
@@ -1,1 +1,0 @@
-"""This package contains legacy end to end tests"""