+++ test/test_perfs.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-import memunit, pytest
+import pytest
 
 from fpdf import FPDF
 
@@ -8,8 +8,10 @@ HERE = Path(__file__).resolve().parent
 
 
 @pytest.mark.timeout(40)
-# ensure memory usage does not get too high - this value depends on Python version:
-@memunit.assert_lt_mb(185)
+# Note: this does not combine well with memunit.
+# memory_profiler.MemTimer does not terminate properly when a signal is raised by pytest-timeout,
+# and as a consequence multiprocessing.util._exit_function becomes blocking at the end of Pytest execution,
+# (on line "calling join() for process MemTimer-1")
 def test_intense_image_rendering():
     png_file_paths = []
     for png_file_path in (HERE / "image/png_images/").glob("*.png"):